python-backend/pyproject.toml
[tool.poetry]
name = "local-talent-py"
version = "0.1.0"
description = "FastAPI backend for LocalTalent (mirrors Node API)"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
asyncpg = "^0.27.0"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
start = "uvicorn main:app --host 0.0.0.0 --port ${PORT:-8000}"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"